Move getting UnflaggedTopics to Route Data Loader
Hey Matt,

Not sure if you have seen this yet but the Router has a nice set of methods you can use to load data and display loading text/images. It does it through a promise so the page will load and the data will load asynchronous. See more here:  http://vuejs.github.io/vue-router/en/pipeline/hooks.html

Will give the user a better experience if the http response is slow or allot of data  :-)